+++ py-polars/tests/unit/test_streaming.py
@@ -317,6 +317,10 @@ def test_streaming_apply(monkeypatch: Any, capfd: Any) -> None:
     monkeypatch.setenv("POLARS_VERBOSE", "1")
     q = pl.DataFrame({"a": [1, 2]}).lazy()
 
-    (q.select(pl.col("a").apply(lambda x: x * 2)).collect(streaming=True))
+    (
+        q.select(pl.col("a").apply(lambda x: x * 2, return_dtype=pl.Int64)).collect(
+            streaming=True
+        )
+    )
     (_, err) = capfd.readouterr()
     assert "df -> projection -> ordered_sink" in err
